2023-07-18 Fixing error `class java.lang.Integer cannot be cast to class java.lang.Long` in org.apache.kafka.common.config.AbstractConfig.getLong(AbstractConfig.java:181)

+++ src/kafka/kafka-key-value-storage/src/main/java/dev/iakunin/library/kafka/keyvaluestorage/service/TopicInitializer.java
@@ -96,7 +96,7 @@ public class TopicInitializer implements Consumer<TopicAdmin> {
         return createTopicsWithRetry(
             admin,
             topicDescription,
-            adminClientConfig.getLong(AdminClientConfig.DEFAULT_API_TIMEOUT_MS_CONFIG),
+            (long) adminClientConfig.getInt(AdminClientConfig.DEFAULT_API_TIMEOUT_MS_CONFIG),
             adminClientConfig.getLong(AdminClientConfig.RETRY_BACKOFF_MS_CONFIG),
             time
         );
